finish

+++ python_api/SparkVersion/Operations/ReadFileOperation.py
@@ -4,11 +4,12 @@ from SparkVersion.Operations.Helpers.StatisticsCreator import createStatisticsSu
 
 
 def readFile(filename):
-    file = Context.spark.read\
-      .option("inferSchema", "true")\
-      .option("header", "true")\
-      .option("sep", ",")\
-      .csv(Context.defaultDirectoryPath + filename)\
-      .na.drop()
+    file = Context.spark.read \
+        .option("inferSchema", "true") \
+        .option("header", "true") \
+        .option("sep", ",") \
+        .csv(Context.defaultDirectoryPath + filename) \
+        .na.drop() if filename.endswith(".csv") else \
+        Context.spark.read.option("inferSchema", "true").parquet(Context.defaultDirectoryPath + filename).na.drop()
     safeData(file)
     return createStatisticsSummary(file)
